Add controls columns to <FilterTable />
This type of column is inteded to hold actions to be taken over
the row data which the button is placed with a callback function
to perform the action with the relevant data passed as argument.

Make use of it in <MainContainer /> to add ability to show/hide
markers by clicking in the buttons.